Remove Scroll Locking Feature
Limiting the body, html, .ddm-menu-container, .ddm-menu-container__inner
to 100% height and turning off scrolling on them is causing too many
issues. These features are not worth the pain.

Some of the known issues caused are as follows:

* Cannot scroll with keyboard
* Mobile Chrome does not hide on scroll
* Cannot click on the top of a mobile browser to jump back to top
* Cannot scroll at all in Android 2
* Cannot scroll on ad wings on KSL Homes
* Positioning of wings on KSL Homes is difficult as the absolute
  position right and left values depend on whether the inner scroll
  bar shows or does not show
* Absolutely positioned content can show on top of the scroll bars.

To solve these issues, I've removed the scroll lock issue. We can
reimplement it later once a good solution is found but it's becoming
too painful to wait for a fix.